+++ src/ralph/models/xapi/__init__.py
@@ -1,0 +1,15 @@
+"""xAPI pydantic models"""
+
+# flake8: noqa
+
+from .video.statements import (
+    VideoCompleted,
+    VideoInitialized,
+    VideoInteracted,
+    VideoPaused,
+    VideoPlayed,
+    VideoSeeked,
+    VideoTerminated,
+)
+
+from .navigation.statements import PageTerminated, PageViewed