+++ src/ralph/models/xapi/navigation/statements.py
@@ -1,6 +1,8 @@
 """Navigation xAPI event definitions"""
 
+from ...selector import selector
 from ..base import BaseXapiModel
+from ..constants import ACTIVITY_PAGE_ID, VERB_TERMINATED_ID, VERB_VIEWED_ID
 from ..fields.verbs import TerminatedVerbField, ViewedVerbField
 from .fields.objects import PageObjectField
 
@@ -15,6 +17,11 @@ class PageViewed(BaseXapiModel):
        verb (PageViewedVerbField): See PageViewedVerbField.
     """
 
+    __selector__ = selector(
+        object__definition__type="http://activitystrea.ms/schema/1.0/page",
+        verb__id="http://id.tincanapi.com/verb/viewed",
+    )
+
     object: PageObjectField
     verb: ViewedVerbField = ViewedVerbField()
 
@@ -28,6 +35,11 @@ class PageTerminated(BaseXapiModel):
        object (PageObjectField): See PageObjectField.
        verb (PageTerminatedVerbField): See PageTerminatedVerbField.
     """
+
+    __selector__ = selector(
+        object__definition__type="http://activitystrea.ms/schema/1.0/page",
+        verb__id="http://adlnet.gov/expapi/verbs/terminated",
+    )
 
     object: PageObjectField
     verb: TerminatedVerbField = TerminatedVerbField()