+++ tests/models/xapi/test_navigation.py
@@ -2,10 +2,12 @@
 
 from hypothesis import given, provisional, settings
 from hypothesis import strategies as st
+import json
 
 from ralph.models.xapi.fields.actors import ActorAccountField, ActorField
 from ralph.models.xapi.navigation.fields.objects import PageObjectField
 from ralph.models.xapi.navigation.statements import PageTerminated, PageViewed
+from ralph.models.selector import ModelSelector
 
 
 @settings(max_examples=1)
@@ -46,3 +48,23 @@ def test_models_xapi_page_viewed_statement(statement):
 
     assert statement.verb.id == "http://id.tincanapi.com/verb/viewed"
     assert statement.object.definition.type == "http://activitystrea.ms/schema/1.0/page"
+
+
+@settings(max_examples=1)
+@given(
+    st.builds(
+        PageViewed,
+        actor=st.builds(
+            ActorField,
+            account=st.builds(
+                ActorAccountField, name=st.just("username"), homePage=provisional.urls()
+            ),
+        ),
+        object=st.builds(PageObjectField, id=provisional.urls()),
+    )
+)
+def test_models_xapi_page_viewed_selector_with_valid_statement(statement):
+    """Tests given a page viewed statement, the get_model method should return PageViewed model."""
+
+    statement = json.loads(statement.json())
+    assert ModelSelector(module="ralph.models.xapi").get_model(statement) is PageViewed